+++ pystac/item.py
@@ -87,10 +87,10 @@ class CommonMetadata:
         else:
             start_datetime = asset.properties.get("start_datetime")
 
-        if start_datetime:
-            start_datetime = str_to_datetime(start_datetime)
+        if start_datetime is None:
+            return None
 
-        return start_datetime
+        return str_to_datetime(start_datetime)
 
     def set_start_datetime(
         self, start_datetime: Optional[Datetime], asset: Optional[Asset] = None
@@ -138,10 +138,10 @@ class CommonMetadata:
         else:
             end_datetime = asset.properties.get("end_datetime")
 
-        if end_datetime:
-            end_datetime = str_to_datetime(end_datetime)
+        if end_datetime is None:
+            return None
 
-        return end_datetime
+        return str_to_datetime(end_datetime)
 
     def set_end_datetime(
         self, end_datetime: Optional[Datetime], asset: Optional[Asset] = None
@@ -231,10 +231,10 @@ class CommonMetadata:
         else:
             providers = asset.properties.get("providers")
 
-        if providers is not None:
-            providers = [Provider.from_dict(d) for d in providers]
+        if providers is None:
+            return None
 
-        return providers
+        return [Provider.from_dict(d) for d in providers]
 
     def set_providers(
         self, providers: Optional[List[Provider]], asset: Optional[Asset] = None
@@ -494,10 +494,10 @@ class CommonMetadata:
         else:
             created = asset.properties.get("created")
 
-        if created:
-            created = str_to_datetime(created)
+        if created is None:
+            return None
 
-        return created
+        return str_to_datetime(created)
 
     def set_created(
         self, created: Optional[Datetime], asset: Optional[Asset] = None
@@ -560,10 +560,10 @@ class CommonMetadata:
         else:
             updated = asset.properties.get("updated")
 
-        if updated:
-            updated = str_to_datetime(updated)
+        if updated is None:
+            return None
 
-        return updated
+        return str_to_datetime(updated)
 
     def set_updated(
         self, updated: Optional[Datetime], asset: Optional[Asset] = None