+++ pystac/serialization/common_properties.py
@@ -36,10 +36,7 @@ def merge_common_properties(
 
     # The commons extension was removed in 1.0.0-beta.1, so if this is an earlier STAC
     # item we don't have to bother with merging.
-    if (
-        stac_version is not None
-        and STACVersionID(stac_version) > "0.9.0"  # type:ignore
-    ):
+    if stac_version is not None and STACVersionID(stac_version) > "0.9.0":
         return False
 
     # Check to see if this is a 0.9.0 item that
@@ -71,7 +68,7 @@ def merge_common_properties(
             (link for link in links if link["rel"] == pystac.RelType.COLLECTION), None
         )
         if collection_link is not None:
-            collection_href = cast(Dict[str, Any], collection_link).get("href")
+            collection_href = collection_link.get("href")
             if collection_href is not None:
                 if json_href is not None:
                     collection_href = make_absolute_href(collection_href, json_href)