+++ pystac/stac_io.py
@@ -14,16 +14,16 @@ from typing import (
 )
 import warnings
 
-from urllib.parse import urlparse
 from urllib.request import urlopen
 from urllib.error import HTTPError
 
 import pystac
+from pystac.utils import safe_urlparse
 import pystac.serialization
 
 # Use orjson if available
 try:
-    import orjson  # type: ignore
+    import orjson
 except ImportError:
     orjson = None  # type: ignore[assignment]
 
@@ -74,10 +74,12 @@ class StacIO(ABC):
         raise NotImplementedError("write_text not implemented")
 
     def _json_loads(self, txt: str, source: Union[str, "Link_Type"]) -> Dict[str, Any]:
+        result: Dict[str, Any]
         if orjson is not None:
-            return orjson.loads(txt)
+            result = orjson.loads(txt)
         else:
-            return json.loads(txt)
+            result = json.loads(txt)
+        return result
 
     def _json_dumps(
         self, json_dict: Dict[str, Any], source: Union[str, "Link_Type"]
@@ -180,16 +182,18 @@ class DefaultStacIO(StacIO):
         return self.read_text_from_href(href, *args, **kwargs)
 
     def read_text_from_href(self, href: str, *args: Any, **kwargs: Any) -> str:
-        parsed = urlparse(href)
+        parsed = safe_urlparse(href)
+        href_contents: str
         if parsed.scheme != "":
             try:
                 with urlopen(href) as f:
-                    return f.read().decode("utf-8")
+                    href_contents = f.read().decode("utf-8")
             except HTTPError as e:
                 raise Exception("Could not read uri {}".format(href)) from e
         else:
-            with open(href) as f:
-                return f.read()
+            with open(href, encoding="utf-8") as f:
+                href_contents = f.read()
+        return href_contents
 
     def write_text(
         self, dest: Union[str, "Link_Type"], txt: str, *args: Any, **kwargs: Any
@@ -209,7 +213,7 @@ class DefaultStacIO(StacIO):
         dirname = os.path.dirname(href)
         if dirname != "" and not os.path.isdir(dirname):
             os.makedirs(dirname)
-        with open(href, "w") as f:
+        with open(href, "w", encoding="utf-8") as f:
             f.write(txt)
 
 
@@ -225,9 +229,10 @@ class DuplicateKeyReportingMixin(StacIO):
     """
 
     def _json_loads(self, txt: str, source: Union[str, "Link_Type"]) -> Dict[str, Any]:
-        return json.loads(
+        result: Dict[str, Any] = json.loads(
             txt, object_pairs_hook=self.duplicate_object_names_report_builder(source)
         )
+        return result
 
     @staticmethod
     def duplicate_object_names_report_builder(
@@ -349,7 +354,8 @@ class STAC_IO:
             STAC_IO in order to enable additional URI types, replace that member
             with your own implementation.
         """
-        return json.loads(STAC_IO.read_text(uri))
+        result: Dict[str, Any] = json.loads(STAC_IO.read_text(uri))
+        return result
 
     @classmethod
     def read_stac_object(