+++ tests/test_item.py
@@ -3,7 +3,6 @@ from datetime import datetime
 import json
 from typing import Any, Dict, List
 import unittest
-from tempfile import TemporaryDirectory
 
 import pystac
 from pystac import Asset, Item, Provider
@@ -11,14 +10,14 @@ from pystac.validation import validate_dict
 import pystac.serialization.common_properties
 from pystac.item import CommonMetadata
 from pystac.utils import datetime_to_str, get_opt, str_to_datetime, is_absolute_href
-from tests.utils import TestCases, assert_to_from_dict
+from tests.utils import TestCases, assert_to_from_dict, get_temp_dir
 
 
 class ItemTest(unittest.TestCase):
     def get_example_item_dict(self) -> Dict[str, Any]:
         m = TestCases.get_path("data-files/item/sample-item.json")
         with open(m) as f:
-            item_dict = json.load(f)
+            item_dict: Dict[str, Any] = json.load(f)
         return item_dict
 
     def test_to_from_dict(self) -> None:
@@ -73,7 +72,7 @@ class ItemTest(unittest.TestCase):
 
         item.extra_fields["test"] = "extra"
 
-        with TemporaryDirectory() as tmp_dir:
+        with get_temp_dir() as tmp_dir:
             p = os.path.join(tmp_dir, "item.json")
             item.save_object(include_self_link=False, dest_href=p)
             with open(p) as f: